CSP-673 created MyApprenticeshipsController and updated TrainingProviderId type

+++ src/SFA.DAS.ApprenticeAccounts/Application/Commands/CreateMyApprenticeCommand/CreateMyApprenticeshipCommand.cs
@@ -12,7 +12,7 @@ public class CreateMyApprenticeshipCommand : IMyApprenticeshipCommand, IRequest<
     public string? EmployerName { get; set; }
     public DateTime? StartDate { get; set; }
     public DateTime? EndDate { get; set; }
-    public int? TrainingProviderId { get; set; }
+    public long? TrainingProviderId { get; set; }
     public string? TrainingProviderName { get; set; }
     public string? TrainingCode { get; set; }
     public string? StandardUId { get; set; }
